Move fit_linear_probe to the base class, add test labels support
